refactor: split out integration tests to separate classes. Extracted constants to interface. Common test methods are now in intbasetest
