+++ am-lib/src/test/java/integration/uk/gov/hmcts/reform/amlib/GrantAccessIntegrationTest.java
@@ -1,0 +1,27 @@
+package integration.uk.gov.hmcts.reform.amlib;
+
+import org.junit.Before;
+import org.junit.Test;
+
+import java.util.UUID;
+
+import static integration.uk.gov.hmcts.reform.amlib.Constants.ACCESSOR_ID;
+import static integration.uk.gov.hmcts.reform.amlib.Constants.EXPLICIT_READ_CREATE_UPDATE_PERMISSIONS;
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class GrantAccessIntegrationTest extends IntegrationBaseTest {
+
+    private String resourceId;
+
+    @Before
+    public void setupTest() {
+        resourceId = UUID.randomUUID().toString();
+    }
+
+    @Test
+    public void createQuery_whenCreatingResourceAccess_ResourceAccessAppearsInDatabase() {
+        ams.createResourceAccess(createRecord(resourceId, ACCESSOR_ID, EXPLICIT_READ_CREATE_UPDATE_PERMISSIONS));
+
+        assertThat(countResourcesById(resourceId)).isEqualTo(1);
+    }
+}
